Codechange: remove vim foldmarkers
Instead of using fold markers that only vim supports to make
the code more readable, a better choice would be to move classes
in their own files in a new folder. This means other editors
also benefit from the better clustering of information.

As currently the markers drifted from their intended purpose,
it is better to remove them and leave the splitting up for an
day.